This branch revisits the idea of how to compare strings across training and testing.
In brief, as we build a dictionary over training trials, we keep all common strings across trials (and keep the frequency with which they show up) and not just the longest string.

While building the dictionary, we also do a single pass comparing the 0th trial to all the i trials. But we do not do multiple passes, where when we are in an ith+k trial, we look back for possible matches. Here we just will give up on some possible extra combinations and hope the averaging of folds at the end, still is good.

In this branch we also will test the longest common sequence technique besides the longest common subsequence. This new sub-module is inside lcs.LCS